Add files via upload
This is the main code that can be used to perform data augmentation

The highest count of images among our categories is 2749 (Paper). We are setting this as the target count for balancing. This approach would require augmenting images in the under-represented categories.
Here under-represented categories are Wood - 347 medical -1605 cardboard - 2332, we used augmentation techniques (rotation, flipping, zooming, etc.) to generate new samples and balance them.